Better error handling and timeout control.
If there is some sort of connectivity issue with a provider it would
likely just crash the cart. I think it would be better to just use
the fallback price and allow them to continue.

Also one of the errors could be a Net::HTTP timeout which defaults
to 30 seconds. This is way too long to find out things have gone
wrong. So surrounding API interaction with a 5 second timeout to
keep things more reasonable.