Better insurance support.
Previous if using USPS the package was insured for the shipment value.
But this may be wrong as perhaps we are replacing a small cheap part
on the helmet a customer sent in. The entire expensive helmet is in the
package but the customer wasn't charged for their own helmet.

With this new setup the insurance defaults to the shipment value still
but the user can edit it before generating the label allowing that value
to be manually adjusted.

In addition the UPS provider has been monkey-patched to support an insurance
value so both providers are treated the same.