Allow insurance to be enabled/disabled.
At the model level we have two attributes. The insurance amount as well as
if insurance is enabled/disabled. This allows us to disable the insurance but
still keep the old value of the insurance should we decide to enable it again.

Tests are a bit light for two reasons:

* It interacts with 3rd party services and mocks/stubs would have dubious value.
* It's only a fragment of the UI that is fully implemented in the application.