Present label error in UI.
Previously we just had a generic error regardless of the cause. I didn't
think it would be a common occurance so I just `!` it and let it bubble up
as a 500 error.

In reality labels fail to generate for a variety of reasons. Some may be due
to infrasturcture reasons (invalid key, networking issues, etc) while others
may be due to transactional reasons (not enough money in provide account,
invalid address).

By bubbling up the true error we let the user correct those transactional issues
themselves reducing support on our side and letting them correct the issue
more timely. Even when it is infrastructural if we get a screenshot then we will
have better information to start with and not have to dig through the logs.